+++ op-deployer/pkg/deployer/opcm/implementations.go
@@ -1,12 +1,10 @@
 package opcm
 
 import (
-	"fmt"
 	"math/big"
 
-	"github.com/ethereum/go-ethereum/common"
-
 	"github.com/ethereum-optimism/optimism/op-chain-ops/script"
+	"github.com/ethereum/go-ethereum/common"
 )
 
 type DeployImplementationsInput struct {
@@ -22,11 +20,6 @@ type DeployImplementationsInput struct {
 	ProtocolVersionsProxy common.Address
 	SuperchainProxyAdmin  common.Address
 	UpgradeController     common.Address
-	UseInterop            bool // if true, deploy Interop implementations
-}
-
-func (input *DeployImplementationsInput) InputSet() bool {
-	return true
 }
 
 type DeployImplementationsOutput struct {
@@ -38,7 +31,7 @@ type DeployImplementationsOutput struct {
 	OpcmInteropMigrator              common.Address `json:"opcmInteropMigratorAddress"`
 	DelayedWETHImpl                  common.Address `json:"delayedWETHImplAddress"`
 	OptimismPortalImpl               common.Address `json:"optimismPortalImplAddress"`
-	ETHLockboxImpl                   common.Address `json:"ethLockboxImplAddress" evm:"ethLockboxImpl"`
+	ETHLockboxImpl                   common.Address `json:"ethLockboxImplAddress" abi:"ethLockboxImpl"`
 	PreimageOracleSingleton          common.Address `json:"preimageOracleSingletonAddress"`
 	MipsSingleton                    common.Address `json:"mipsSingletonAddress"`
 	SystemConfigImpl                 common.Address `json:"systemConfigImplAddress"`
@@ -52,56 +45,9 @@ type DeployImplementationsOutput struct {
 	ProtocolVersionsImpl             common.Address `json:"protocolVersionsImplAddress"`
 }
 
-func (output *DeployImplementationsOutput) CheckOutput(input common.Address) error {
-	return nil
-}
+type DeployImplementationsScript script.DeployScriptWithOutput[DeployImplementationsInput, DeployImplementationsOutput]
 
-type DeployImplementationsScript struct {
-	Run func(input, output common.Address) error
-}
-
-func DeployImplementations(
-	host *script.Host,
-	input DeployImplementationsInput,
-) (DeployImplementationsOutput, error) {
-	var output DeployImplementationsOutput
-	inputAddr := host.NewScriptAddress()
-	outputAddr := host.NewScriptAddress()
-
-	cleanupInput, err := script.WithPrecompileAtAddress[*DeployImplementationsInput](host, inputAddr, &input)
-	if err != nil {
-		return output, fmt.Errorf("failed to insert DeployImplementationsInput precompile: %w", err)
-	}
-	defer cleanupInput()
-
-	cleanupOutput, err := script.WithPrecompileAtAddress[*DeployImplementationsOutput](host, outputAddr, &output,
-		script.WithFieldSetter[*DeployImplementationsOutput])
-	if err != nil {
-		return output, fmt.Errorf("failed to insert DeployImplementationsOutput precompile: %w", err)
-	}
-	defer cleanupOutput()
-
-	implContract := "DeployImplementations"
-	deployScript, cleanupDeploy, err := script.WithScript[DeployImplementationsScript](host, "DeployImplementations.s.sol", implContract)
-	if err != nil {
-		return output, fmt.Errorf("failed to load %s script: %w", implContract, err)
-	}
-	defer cleanupDeploy()
-
-	opcmContract := "OPContractsManager"
-	if err := host.RememberOnLabel("OPContractsManager", opcmContract+".sol", opcmContract); err != nil {
-		return output, fmt.Errorf("failed to link OPContractsManager label: %w", err)
-	}
-
-	// So we can see in detail where the SystemConfig interop initializer fails
-	sysConfig := "SystemConfig"
-	if err := host.RememberOnLabel("SystemConfigImpl", sysConfig+".sol", sysConfig); err != nil {
-		return output, fmt.Errorf("failed to link SystemConfig label: %w", err)
-	}
-
-	if err := deployScript.Run(inputAddr, outputAddr); err != nil {
-		return output, fmt.Errorf("failed to run %s script: %w", implContract, err)
-	}
-
-	return output, nil
+// NewDeployImplementationsScript loads and validates the DeployImplementations script contract
+func NewDeployImplementationsScript(host *script.Host) (DeployImplementationsScript, error) {
+	return script.NewDeployScriptWithOutputFromFile[DeployImplementationsInput, DeployImplementationsOutput](host, "DeployImplementations.s.sol", "DeployImplementations")
 }
