+++ op-deployer/pkg/deployer/pipeline/implementations.go
@@ -45,8 +45,11 @@ func DeployImplementations(env *Env, intent *state.Intent, st *state.State) erro
 		return fmt.Errorf("error merging proof params from overrides: %w", err)
 	}
 
-	dio, err := opcm.DeployImplementations(
-		env.L1ScriptHost,
+	deployScript, err := opcm.NewDeployImplementationsScript(env.L1ScriptHost)
+	if err != nil {
+		return fmt.Errorf("failed to load DeployImplementations script: %w", err)
+	}
+	dio, err := deployScript.Run(
 		opcm.DeployImplementationsInput{
 			WithdrawalDelaySeconds:          new(big.Int).SetUint64(proofParams.WithdrawalDelaySeconds),
 			MinProposalSizeBytes:            new(big.Int).SetUint64(proofParams.MinProposalSizeBytes),
@@ -59,7 +62,6 @@ func DeployImplementations(env *Env, intent *state.Intent, st *state.State) erro
 			ProtocolVersionsProxy:           st.SuperchainDeployment.ProtocolVersionsProxy,
 			SuperchainProxyAdmin:            st.SuperchainDeployment.SuperchainProxyAdminImpl,
 			UpgradeController:               st.SuperchainRoles.SuperchainProxyAdminOwner,
-			UseInterop:                      intent.UseInterop,
 		},
 	)
 	if err != nil {
